+++ Fire.h
@@ -11,7 +11,7 @@ class Fire
 {
   public:
     Fire();
-	Fire(int);	// Timer with set milliseconds
+	Fire(unsigned long);	// Timer with set milliseconds
 	void setCallback(void (*callback_function) () );
 	bool fire();
 	bool fireExpired();
@@ -19,7 +19,7 @@ class Fire
 	bool updateInterval(int);
 	
   private:
-	int _fireInterval = 1000;	// 1 second by default.
+	unsigned long _fireInterval = 1000;	// 1 second by default.
 	void (*_callback)();
 	bool _isCallbackEnabled = false;
 };
